IMHO, "reconnect" should catch __ALL__ sorts of connection errors.
Rationale: I'm using this stuff for a rather primitive bot which is tied
to a virtual snd_aloop device and just forwards anything to the mumble
server in the background -- and receives any input from a mumble server
in the background and forwards this "noise" to another Alsa loopback
channel.

As background process it is meant to run standalone, un-supervised and
stable.

Without this patch pymumble will just bail-out on simple network errors,
or when the mumble server is out-for-lunch-in-maintenance-mode.

IMHO, pymumble should just be penetrant and insistant with
"reconnect=True" and just should not bail out. Simply retry forever.

Other thoughts: the behaviour implemented by this patch could also be
achieved by the "calling application" by catching the asserted
exception. However, if a simple network error also bails out even with
"reconnect=True", then the purpose of "reconnect=True" is no longer
clear to me.

No flames intented, and kind thanks for your good work!